+++ CHANGELOG.md
@@ -12,9 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Do not fetch diff updates until the next full update, if some fatal error occurred during the last update [#2717].
+- Check filter checksum after it was downloaded [#2681].
 
 [Unreleased]: https://github.com/AdguardTeam/AdguardBrowserExtension/compare/v4.3.35...HEAD
 [#2717]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2717
+[#2681]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2681
 
 
 <!-- TODO: add an actual date after github release -->